@@ -1,31 +1,99 @@
-﻿using Serilog.Configuration;
+﻿using System.Collections.Generic;
+using Serilog.Configuration;
 using Serilog.Core;
 using Serilog.Events;
 using Serilog.Sinks.TestCorrelator;
 
-namespace Serilog
+namespace Serilog;
+
+/// <summary>
+/// Adds the WriteTo.TestCorrelator extension methods to <seealso cref="LoggerSinkConfiguration"/>.
+/// </summary>
+public static class LoggerSinkConfigurationExtensions
 {
     /// <summary>
-    /// Adds the WriteTo.TestCorrelator extension method to LoggerSinkConfiguration.
+    /// Writes <seealso cref="LogEvent"/>s to a <see cref="TestCorrelatorSink"/>.
     /// </summary>
-    public static class LoggerSinkConfigurationExtensions
-    {
-        /// <summary>
-        /// Adds a <see cref="TestCorrelatorSink"/> to write to.
-        /// </summary>
-        /// <param name="loggerSinkConfiguration">The LoggerSinkConfiguration to write to.</param>
-        /// <param name="restrictedToMinimumLevel">The minimum LogEventLevel for events passed to the sink. Ignored when <paramref name="levelSwitch"/> is specified.</param>
-        /// <param name="levelSwitch">A switch allowing the minimum LogEventLevel to be changed at runtime.</param>
-        /// <returns>A modified LoggerConfiguration allowing configuration to continue.</returns>
-        public static LoggerConfiguration TestCorrelator(
-            this LoggerSinkConfiguration loggerSinkConfiguration,
-            LogEventLevel restrictedToMinimumLevel = LevelAlias.Minimum,
-            LoggingLevelSwitch levelSwitch = null)
-        {
-            return loggerSinkConfiguration.Sink(
-                new TestCorrelatorSink(),
-                restrictedToMinimumLevel,
-                levelSwitch);
-        }
-    }
+    /// <param name="loggerSinkConfiguration">The <seealso cref="LoggerSinkConfiguration"/> to write to.</param>
+    /// <param name="ids">The <seealso cref="TestCorrelatorSinkId"/>s to apply to this sink.</param>
+    /// <returns>A modified <seealso cref="LoggerConfiguration"/> allowing configuration to continue.</returns>
+    public static LoggerConfiguration TestCorrelator(
+        this LoggerSinkConfiguration loggerSinkConfiguration,
+        params TestCorrelatorSinkId[] ids) =>
+        loggerSinkConfiguration.Sink(new TestCorrelatorSink(ids));
+
+    /// <summary>
+    /// Writes <seealso cref="LogEvent"/>s to a <see cref="TestCorrelatorSink"/>.
+    /// </summary>
+    /// <param name="loggerSinkConfiguration">The <seealso cref="LoggerSinkConfiguration"/> to write to.</param>
+    /// <param name="restrictedToMinimumLevel">The minimum <seealso cref="LogEventLevel"/> for events passed to the sink.</param>
+    /// <param name="ids">The <seealso cref="TestCorrelatorSinkId"/>s to apply to this sink.</param>
+    /// <returns>A modified <seealso cref="LoggerConfiguration"/> allowing configuration to continue.</returns>
+    public static LoggerConfiguration TestCorrelator(
+        this LoggerSinkConfiguration loggerSinkConfiguration,
+        LogEventLevel restrictedToMinimumLevel,
+        params TestCorrelatorSinkId[] ids) =>
+        loggerSinkConfiguration.Sink(
+            new TestCorrelatorSink(ids),
+            restrictedToMinimumLevel,
+            null);
+
+    /// <summary>
+    /// Writes <seealso cref="LogEvent"/>s to a <see cref="TestCorrelatorSink"/>.
+    /// </summary>
+    /// <param name="loggerSinkConfiguration">The <seealso cref="LoggerSinkConfiguration"/> to write to.</param>
+    /// <param name="levelSwitch">A <seealso cref="LoggingLevelSwitch"/> allowing the minimum <seealso cref="LogEventLevel"/> to be changed at runtime.</param>
+    /// <param name="ids">The <seealso cref="TestCorrelatorSinkId"/>s to apply to this sink.</param>
+    /// <returns>A modified <seealso cref="LoggerConfiguration"/> allowing configuration to continue.</returns>
+    public static LoggerConfiguration TestCorrelator(
+        this LoggerSinkConfiguration loggerSinkConfiguration,
+        LoggingLevelSwitch levelSwitch,
+        params TestCorrelatorSinkId[] ids) =>
+        loggerSinkConfiguration.Sink(
+            new TestCorrelatorSink(ids),
+            LevelAlias.Minimum,
+            levelSwitch);
+
+    /// <summary>
+    /// Writes <seealso cref="LogEvent"/>s to a <see cref="TestCorrelatorSink"/>.
+    /// </summary>
+    /// <param name="loggerSinkConfiguration">The <seealso cref="LoggerSinkConfiguration"/> to write to.</param>
+    /// <param name="ids">The <seealso cref="TestCorrelatorSinkId"/>s to apply to this sink.</param>
+    /// <returns>A modified <seealso cref="LoggerConfiguration"/> allowing configuration to continue.</returns>
+    public static LoggerConfiguration TestCorrelator(
+        this LoggerSinkConfiguration loggerSinkConfiguration,
+        IEnumerable<TestCorrelatorSinkId> ids) =>
+        loggerSinkConfiguration.Sink(new TestCorrelatorSink(ids));
+
+    /// <summary>
+    /// Writes <seealso cref="LogEvent"/>s to a <see cref="TestCorrelatorSink"/>.
+    /// </summary>
+    /// <param name="loggerSinkConfiguration">The <seealso cref="LoggerSinkConfiguration"/> to write to.</param>
+    /// <param name="restrictedToMinimumLevel">The minimum <seealso cref="LogEventLevel"/> for events passed to the sink.</param>
+    /// <param name="ids">The <seealso cref="TestCorrelatorSinkId"/>s to apply to this sink.</param>
+    /// <returns>A modified <seealso cref="LoggerConfiguration"/> allowing configuration to continue.</returns>
+    public static LoggerConfiguration TestCorrelator(
+        this LoggerSinkConfiguration loggerSinkConfiguration,
+        LogEventLevel restrictedToMinimumLevel,
+        IEnumerable<TestCorrelatorSinkId> ids) =>
+        loggerSinkConfiguration.Sink(
+            new TestCorrelatorSink(ids),
+            restrictedToMinimumLevel,
+            null);
+
+    /// <summary>
+    /// Writes <seealso cref="LogEvent"/>s to a <see cref="TestCorrelatorSink"/>.
+    /// </summary>
+    /// <param name="loggerSinkConfiguration">The <seealso cref="LoggerSinkConfiguration"/> to write to.</param>
+    /// <param name="levelSwitch">A <seealso cref="LoggingLevelSwitch"/> allowing the minimum <seealso cref="LogEventLevel"/> to be changed at runtime.</param>
+    /// <param name="ids">The <seealso cref="TestCorrelatorSinkId"/>s to apply to this sink.</param>
+    /// <returns>A modified <seealso cref="LoggerConfiguration"/> allowing configuration to continue.</returns>
+    public static LoggerConfiguration TestCorrelator(
+        this LoggerSinkConfiguration loggerSinkConfiguration,
+        LoggingLevelSwitch levelSwitch,
+        IEnumerable<TestCorrelatorSinkId> ids) =>
+        loggerSinkConfiguration.Sink(
+            new TestCorrelatorSink(ids),
+            LevelAlias.Minimum,
+            levelSwitch);
 }
