@@ -1,22 +1,28 @@
-﻿using FluentAssertions;
+﻿using System;
+using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Serilog.Events;
 
-namespace Serilog.Sinks.TestCorrelator.Tests
+namespace Serilog.Sinks.TestCorrelator.Tests;
+
+[TestClass]
+public class TestCorrelatorSinkTests
 {
-    [TestClass]
-    public class TestCorrelatorSinkTests
+    [TestMethod]
+    public void A_TestCorrelatorSink_writes_LogEvents_emitted_to_it_to_a_TestCorrelator()
     {
-        [TestMethod]
-        public void A_TestCorrelatorSink_writes_LogEvents_emited_to_it_to_a_TestCorrelator()
+        LogEvent logEvent = new(
+            DateTimeOffset.UtcNow,
+            LogEventLevel.Information,
+            null,
+            new MessageTemplate([]),
+            []);
+
+        using (TestCorrelator.CreateContext())
         {
-            var logger = new LoggerConfiguration().WriteTo.Sink(new TestCorrelatorSink()).CreateLogger();
+            new TestCorrelatorSink().Emit(logEvent);
 
-            using (TestCorrelator.CreateContext())
-            {
-                logger.Information("");
-
-                TestCorrelator.GetLogEventsFromCurrentContext().Should().ContainSingle();
-            }
+            TestCorrelator.GetLogEventsFromCurrentContext().Should().ContainSingle().Which.Should().Be(logEvent);
         }
     }
 }
